Separate VISH components into their own module, and add warped version
